+++ pkg/github/enterprise/module.go
@@ -1,13 +1,28 @@
-//go:build enterprise
-// +build enterprise
-
 package enterprise
 
 import (
 	"mash/pkg/di"
+	"mash/pkg/github/enterprise/service"
 	"mash/pkg/github/enterprise/workers"
 )
 
 var Module = di.NewModule(
 	di.Needs(workers.Module),
+	di.Needs(service.Module),
+
+	// todo: hack to solve circular dependency
+	di.Provides(func() *service.ImporterQueue {
+		return new(service.ImporterQueue)
+	}),
+	di.Invoke(func(sq *service.ImporterQueue, wq workers.ImporterQueue) {
+		*sq = wq
+	}),
+
+	// todo: hack to solve circular dependency
+	di.Provides(func() *service.ClonerQueue {
+		return new(service.ClonerQueue)
+	}),
+	di.Invoke(func(sq *service.ClonerQueue, wq *workers.ClonerQueue) {
+		*sq = wq
+	}),
 )
