+++ pkg/github/enterprise/routes/webhook.go
@@ -16,8 +16,8 @@ import (
 	"mash/pkg/github/enterprise/routes/push"
 	"mash/pkg/github/enterprise/routes/statuses"
 	"mash/pkg/github/enterprise/routes/workflows"
+	service_github "mash/pkg/github/enterprise/service"
 	workers_github "mash/pkg/github/enterprise/workers"
-	service_github "mash/pkg/github/service"
 	db_review "mash/pkg/review/db"
 	service_statuses "mash/pkg/statuses/service"
 	service_sync "mash/pkg/sync/service"
