+++ pkg/github/enterprise/service/module.go
@@ -1,0 +1,9 @@
+package service
+
+import (
+	"mash/pkg/di"
+)
+
+var Module = di.NewModule(
+	di.Provides(New),
+)
