+++ pkg/github/enterprise/workers/importer.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"time"
 
-	service_github "mash/pkg/github/service"
+	service_github "mash/pkg/github/enterprise/service"
 	"mash/pkg/queue"
 	"mash/pkg/queue/names"
